412 reseting max agents (#441)
* Added new Type for events, added resetting_usage to events and crossings.

* Some minor changes due to code conventions and reordering

+++ math/GradientModel.cpp
@@ -114,7 +114,7 @@ bool GradientModel::Init (Building* building)
      pedsToRemove.clear();
      bool error_occurred = false;
 #pragma omp parallel for
-    for(int p=0;p<allPeds.size();p++) {
+    for(size_t p=0;p<allPeds.size();p++) {
          Pedestrian* ped = allPeds[p];
          double cosPhi = 0;
          double sinPhi = 0;
