+++ routing/ff_router/UnivFFviaFM.cpp
@@ -440,7 +440,7 @@ void UnivFFviaFM::recreateAllForQuickest() {
 #pragma omp parallel
      {
 #pragma omp for
-          for (int i = 0; i < _doors.size(); ++i) {
+          for (size_t i = 0; i < _doors.size(); ++i) {
                auto doorPair = _doors.begin();
                std::advance(doorPair, i);
                addTarget(doorPair->first, _costFieldWithKey[doorPair->first], _directionFieldWithKey[doorPair->first]);
@@ -1456,7 +1456,7 @@ void UnivFFviaFM::addAllTargetsParallel() {
 #pragma omp parallel
      {
 #pragma omp for
-          for (int i = 0; i < _doors.size(); ++i) {
+          for (size_t i = 0; i < _doors.size(); ++i) {
                auto doorPair = _doors.begin();
                std::advance(doorPair, i);
                addTarget(doorPair->first, _costFieldWithKey[doorPair->first], _directionFieldWithKey[doorPair->first]);
@@ -1489,7 +1489,7 @@ void UnivFFviaFM::addTargetsParallel(std::vector<int> wantedDoors) {
 #pragma omp parallel
      {
 #pragma omp for
-          for (int i = 0; i < wantedDoors.size(); ++i) {
+          for (size_t i = 0; i < wantedDoors.size(); ++i) {
                auto doorUID = wantedDoors.begin();
                std::advance(doorUID, i);
                addTarget(*doorUID, _costFieldWithKey[*doorUID], _directionFieldWithKey[*doorUID]);
